+++ src/composables/use-nft-tokens.ts
@@ -9,51 +9,25 @@ import {
 import { config } from '@/config'
 import { TOKEN_TYPES } from '@/enums'
 import { useNetworksStore, useWeb3ProvidersStore } from '@/store'
-import { PageOrder, Signature, BuyParams, FullBookInfo } from '@/types'
+import {
+  PageOrder,
+  Signature,
+  BuyParams,
+  FullBookInfo,
+  Payment,
+  NftMetadata,
+  TokenBaseInfo,
+  TokenFullInfo,
+} from '@/types'
 import { IMarketplace } from '@/types/contracts/MarketPlace'
 import { BN } from '@/utils/math.util'
 import { computed, ref, watch } from 'vue'
 import { useFetch } from '@vueuse/core'
 import { constants } from 'ethers'
 
-export type TokenBaseInfo = {
-  tokenContract: string
-  tokenId: string
-  metadata: NftMetadata
-}
-
-export type TokenFullInfo = TokenBaseInfo & {
-  payment: Payment
-}
-
 type TokensRaw = {
   tokenContract: string
   tokenIds: Array<string>
-}
-
-type NftMetadata = {
-  name: string
-  description: string
-  image: string
-  external_url: string
-}
-
-type Payment = {
-  amount: string
-  book_id: number
-  contract_address: string
-  minted_token_price: string
-  payer_address: string
-  payment_token_price: string
-  purchase_timestamp: string
-  token_id: number
-  erc20_data: {
-    address: string
-    symbol: string
-    name: string
-    decimals: number
-  }
-  type: TOKEN_TYPES
 }
 
 const MAX_METADATA_WAIT_TIME = 8000 // ms
