+++ src/enums/form.enum.ts
@@ -1,0 +1,6 @@
+export enum FORM_STATES {
+  active = 'active',
+  disabled = 'disabled',
+  pending = 'pending',
+  success = 'success',
+}
