+++ src/types/api/book.types.ts
@@ -23,3 +23,25 @@ export type Book = JsonApiRecordBase<'books'> & {
   file: BookFile
   networks: BookNetwork[]
 }
+
+export type FullBookInfo = Book & {
+  tokenContract: string
+  tokenName: string
+  tokenSymbol: string
+  pricePerOneToken: string
+  minNFTFloorPrice: string
+  voucherTokensAmount: string
+  voucherTokenContract: string
+  fundsRecipient: string
+  isNFTBuyable: boolean
+  isDisabled: boolean
+  isVoucherBuyable: boolean
+}
+
+export type BaseBookInfo = Book & {
+  pricePerOneToken: string
+  isDisabled: boolean
+  tokenContract: string
+  tokenName: string
+  tokenSymbol: string
+}
