+++ src/types/index.ts
@@ -7,13 +7,5 @@ export * from './api'
 export * from './chain.types'
 export * from './guide.types'
 export * from './symbols'
-export type { TokenFullInfo, TokenBaseInfo } from '@/composables/use-nft-tokens'
-export type { FullBookInfo, BaseBookInfo } from '@/composables/use-books'
-export type {
-  Signature,
-  BuyParams,
-} from '@/composables/contracts/use-marketplace'
-export type {
-  UnwrapPromiseProperties,
-  UnwrapPromise,
-} from '@/helpers/promise.helpers'
+export * from './nft-tokens.types'
+export * from './marketplace.types'
